Add Hat application list to the project.
Added: Application list component which displays a list of cards
for each HAT application passed to it.

Added: Reusable Card component for the above list. This will also
be used for other similar lists within the app.

Added: A test HAT application object to allow for testing the above
componentss and ensure they are rendering the correct information.

Added: Text strings into the english translations for the HAT
Applications header.

Changed: Updated the AppRouter to hold a new private route which
displays the Application header and list.

Changed: Refactored TileHeader to take the text information from
translations rather than hard-coded strings. This sets it up to be
usable in multiple languages.