Add Application Details action buttons
Added: A new dropdown menu component that shows when the user clicks
the 'more' button on the Application Details screen.

The menu is only available if the application has a 'setup' status.
If not, then a 'close' button is rendered instead which navigates
the user back a screen.

The user has 3 options when the menu is available:
1) To navigate to the app permmissions page.
2) To disable the Hat Application.
3) Cancel / close the menu.