Add page for Hat Tools and display Tool list.
Changed: Updated the tools link in the side-menu to no longer point
to an external link.

Added: A Tools page to be used, which has a tools-header and as well
as a list of the Tools.
Added: English text for the tools page.
Added: New core 'list' scss file for storing css related to all
list class types.