Add Details Action menu to the Tools screen.
Added: New tools component. When a tool is in an 'enabled' state, a menu
will be presented to the user to allow them to disconnect the tool.
If the tool is not enabled, the user can navigate back a screen instead.

Changed: Refactored existing Action component that is used in Application
Details. There was a lot of overlap, so a new re-usable component is
added that can be used across both screens.
Changed: Moved some action-specific styles into a new core scss file.