Feature/dse 146 tools ui (#93)
* Add page for Hat Tools and display Tool list.

Changed: Updated the tools link in the side-menu to no longer point
to an external link.

Added: A Tools page to be used, which has a tools-header and as well
as a list of the Tools.
Added: English text for the tools page.
Added: New core 'list' scss file for storing css related to all
list class types.

* Add page for Hat Tools and display Tool list.

Changed: Updated the tools link in the side-menu to no longer point
to an external link.

Added: A Tools page to be used, which has a tools-header and as well
as a list of the Tools.
Added: English text for the tools page.
Added: New core 'list' scss file for storing css related to all
list class types.

* Add Tool Details route and screen.

Added: New ToolDetails component. It is displayed when one of the
tools in the Tool list is selected.

Changed: Updated header styles to allow for a small margin for tool
screens.
Changed: Updated the service have separate methods for getting all
tools and a single tool by id.

* Add Details Action menu to the Tools screen.

Added: New tools component. When a tool is in an 'enabled' state, a menu
will be presented to the user to allow them to disconnect the tool.
If the tool is not enabled, the user can navigate back a screen instead.

Changed: Refactored existing Action component that is used in Application
Details. There was a lot of overlap, so a new re-usable component is
added that can be used across both screens.
Changed: Moved some action-specific styles into a new core scss file.